+++ src/common/AppHeader.js
@@ -6,7 +6,7 @@ import {
 
 import { Layout, Menu } from 'antd'
 import ProfileArea from './ProfileArea'
-import { ROUTES } from '../constants'
+import { ROUTES, EMPLOYEE_ROLES } from '../constants'
 
 const Header = Layout.Header;
 
@@ -16,24 +16,21 @@ class AppHeader extends Component {
     }
 
     render() {
+        const { currentUser, onSignOut, isAuthenticated, showSignUpModal, showSignInModal } = this.props
+        const name = currentUser && currentUser.firstName ?  `Welcome ${currentUser.firstName} ${currentUser.lastName}` : ""
         let menuItems = []
-        if (this.props.currentUser) {
+        menuItems.push(<Menu.Item key="/"><Link to="/"><span>Home</span></Link></Menu.Item>)
+        if (currentUser) {   
+            if (currentUser.role !== EMPLOYEE_ROLES.employee)
+                menuItems.push(<Menu.Item key={ROUTES.employees}><Link to={ROUTES.employees}><span>Employees</span></Link></Menu.Item>,)
             
-            menuItems = [
-                <Menu.Item key="/"><Link to="/"><span>Home</span></Link></Menu.Item>,
-                <Menu.Item key={ROUTES.employees}><Link to={ROUTES.employees}><span>Employees</span></Link></Menu.Item>,
+            menuItems = menuItems.concat([
                 <Menu.Item key={ROUTES.roster}><Link to={ROUTES.roster}><span>Roster</span></Link></Menu.Item>,
                 <Menu.Item key={ROUTES.availability}><Link to={ROUTES.availability}><span>Availability</span></Link></Menu.Item>,
                 <Menu.Item key={ROUTES.leave}><Link to={ROUTES.leave}><span>Leaves</span></Link></Menu.Item>
-            ]
+            ])
         }
-        else
-            menuItems = [
-                <Menu.Item key="/"><Link to="/"><span>Home</span></Link></Menu.Item>,
-            ]
-        
-        const { currentUser, onSignOut, isAuthenticated, showSignUpModal, showSignInModal } = this.props
-        const name = currentUser && currentUser.firstName ?  `Welcome ${currentUser.firstName} ${currentUser.lastName}` : ""
+    
         
         return (
             <Header className="app-header">
